Mudei o método de gerar indivíduos. Agora eles são parâmetros aleatórios. Fiz um fitness simples.

+++ include/headers.h
@@ -3,7 +3,7 @@
 
 void teste();
 
-void gerar_individuos(float matriz_individuos[][2], short int  m);
+void gerar_individuos(float matriz_individuos[][2], int  m);
 float calcular_erro(float a, float b, float matriz[][2], int n);
 
 #endif
